+++ app/changelog.md
@@ -354,3 +354,4 @@
 - Fixed form wizard view on mobile.
 - Added getMenu function to triMenu that will grab any menu item using its id attribute
 - Changed Todo example to use new triMenu to change it's badge value based on number of todos
+- Added click handler to menus to allow click events as well as state router changes